+++ src/render/AngleRenderer.cpp
@@ -2791,6 +2791,7 @@ void AngleRenderer::drawReferenceImages(const Scene& scene, const Camera& camera
     GLint locOpacity = glGetUniformLocation(m_refImageProgram, "uOpacity");
     GLint locTexture = glGetUniformLocation(m_refImageProgram, "uTexture");
     GLint locAspectScale = glGetUniformLocation(m_refImageProgram, "uAspectScale");
+    GLint locRotation = glGetUniformLocation(m_refImageProgram, "uRotation");
 
     glActiveTexture(GL_TEXTURE0);
     glUniform1i(locTexture, 0);
@@ -2814,6 +2815,9 @@ void AngleRenderer::drawReferenceImages(const Scene& scene, const Camera& camera
             glUniform1i(locPinned2D, 1);
             glUniform2f(locOffset, img.offsetX, img.offsetY);
             glUniform1f(locScale, img.scale);
+            if (locRotation != -1) {
+                glUniform1f(locRotation, -glm::radians(img.rotation));
+            }
 
             float aspectScaleX = imgAspect / screenAspect;
             float aspectScaleY = 1.0f;
@@ -2823,9 +2827,13 @@ void AngleRenderer::drawReferenceImages(const Scene& scene, const Camera& camera
         } else {
             glEnable(GL_DEPTH_TEST);
             glUniform1i(locPinned2D, 0);
+            if (locRotation != -1) {
+                glUniform1f(locRotation, 0.0f);
+            }
             
             glm::mat4 model = glm::mat4(1.0f);
             model = glm::translate(model, glm::vec3(img.offsetX, img.offsetY, 0.0f));
+            model = glm::rotate(model, -glm::radians(img.rotation), glm::vec3(0.0f, 0.0f, 1.0f));
             model = glm::scale(model, glm::vec3(img.scale * imgAspect, img.scale, 1.0f));
             
             glm::mat4 mvp = camera.getProjMatrix() * camera.getViewMatrix() * model;